network: Switch to callback-based model
This isn't entirely neat, but it's a flexible interface which will do
what we need.

There's now callbacks in the network_txn structure, which are used to
request whoever creates the txn to write or read data.

This means that the network service is totally agnostic to the
transaction content itself. For example, we can implement transactions
which stream direct from a file without needing to read it into memory.
Or indeed read and decompress data on the fly.